added dask for parallel computing
