+++ requirements.txt
@@ -1,4 +1,3 @@
-pyspark
 numpy
 pandas
 dask
